MG456 (Miscellaneous code cleanups and fixes) - Got rid of compiling warnings.
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1023 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Services/Drawing/ServerDrawingService.cpp
@@ -119,7 +119,7 @@ MgByteReader* MgServerDrawingService::DescribeDrawing(MgResourceIdentifier* reso
         char* pDest = strrchr(pBuffer, '>');     // NOXLATE
         if (0 != pDest)
         {
-            int nFilteredBytes = (int)(pDest - pBuffer + 1);
+            size_t nFilteredBytes = pDest - pBuffer + 1;
             if (nFilteredBytes < nBytes)
             {
                 nBytes = nFilteredBytes;
